OpenVR devices changes
Changes how OpenVR devices are added. We add using the device index and also listen for the SteamVR Input action event for the input role. We also assign the roles if needed. Controller roles are needed for controller input to differentiate between left and right hands.
Updated input bindings to be mirrored.