Fix mirror IPD in VR.
We render mirrors to a flipped uv.x so that we preserve chirality.
However, we were missing the step to flip the projection matrix. This works fine in desktop because x is usually 0.
In VR, left and right eye has negative projection matrix X, so we had flipped the left and right eye in the shader.
This mostly worked, but meant that we could not correct for the stereo view offset.
Now we flip the projection matrix instead of swithcing eyes.